pyproject.toml
[project]
name = "front-desk-poc"
version = "0.1.0"
description = "BrightWheel AI Front Desk - voice-first assistant for childcare centers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",

    # LLM
    "anthropic>=0.40.0",

    # Voice input
    "deepgram-sdk>=3.0.0",

    # RAG pipeline
    "pymupdf>=1.24.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",

    # Database
    "aiosqlite>=0.20.0",

    # Config
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # HTTP client
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
